# FinancialApp-main/openapi/openapi.yaml
openapi: 3.0.0
paths:
...
  /auth/register:
    post:
      operationId: AuthController_register
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterDto"
      responses:
        "201":
          description: ""
      tags: &a2
        - auth
  /auth/login:
    post:
      operationId: AuthController_login
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginDto"
      responses:
        "201":
          description: ""
      tags: *a2
  /auth/me:
    get:
      operationId: AuthController_me
      parameters: []
      responses:
        "200":
          description: ""
      tags: *a2
      security:
        - bearer: []
  /auth/refresh:
    post:
      operationId: AuthController_refresh
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenDto"
      responses:
        "201":
          description: ""
      tags: *a2
...
  /auth/oauth/token:
    post:
      operationId: AuthController_oauthToken
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OAuthTokenDto"
      responses:
        "201":
          description: ""
      tags: *a2
  /accounts:
    post:
      operationId: AccountsController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountDto"
      responses:
        "201":
          description: ""
      tags: &a3
        - accounts
      security: &a4
        - bearer: []
    get:
      operationId: AccountsController_findMine
      parameters: []
      responses:
        "200":
          description: ""
      tags: *a3
      security: *a4
  /accounts/{id}:
    get:
      operationId: AccountsController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        "200":
          description: ""
      tags: *a3
      security: *a4
  /accounts/{id}/fund:
    post:
      operationId: AccountsController_fundAccount
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FundAccountDto"
      responses:
        "201":
          description: ""
      tags: *a3
      security: *a4
...
  /ledger/account/{accountId}:
    get:
      operationId: LedgerController_findByAccount
      summary: Получить историю транзакций для счета
      parameters:
        - name: accountId
          required: true
          in: path
          schema:
            type: number
      responses:
        "200":
          description: История транзакций
      tags: *a5
      security: *a6
...
  /card-controls/register:
    post:
      operationId: CardControlsController_register
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterCardDto"
      responses:
        "201":
          description: ""
      tags: &a13
        - card-controls
      security: &a14
        - bearer: []
  /card-controls/tokenize:
    post:
      operationId: CardControlsController_tokenize
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TokenizeCardDto"
      responses:
        "201":
          description: ""
      tags: *a13
      security: *a14
  /card-controls/{cardToken}/freeze:
    post:
      operationId: CardControlsController_freeze
      parameters:
        - name: cardToken
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FreezeDto"
      responses:
        "201":
          description: ""
      tags: *a13
      security: *a14
  /card-controls/{cardToken}/unfreeze:
    post:
      operationId: CardControlsController_unfreeze
      parameters:
        - name: cardToken
          required: true
          in: path
          schema:
            type: string
      responses:
        "201":
          description: ""
      tags: *a13
      security: *a14
  /card-controls/{cardToken}/limits:
    patch:
      operationId: CardControlsController_updateLimits
      parameters:
        - name: cardToken
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLimitsDto"
      responses:
        "200":
          description: ""
      tags: *a13
      security: *a14
...
  /kyc/status:
    get:
      operationId: KycController_getStatus
      summary: Get KYC status and documents for current user
      parameters: []
      responses:
        "200":
          description: KYC status and documents
      tags: &a17
        - kyc
      security: &a18
        - bearer: []
    patch:
      operationId: KycController_updateStatus
      summary: Update KYC status (admin only)
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateKycStatusDto"
      responses:
        "200":
          description: KYC status updated
      tags: *a17
      security: *a18
  /kyc/documents:
    get:
      operationId: KycController_getDocuments
      summary: Get all KYC documents for current user
      parameters: []
      responses:
        "200":
          description: List of KYC documents
      tags: *a17
      security: *a18
    post:
      operationId: KycController_submitDocument
      summary: Submit KYC document
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitKycDocumentDto"
      responses:
        "201":
          description: Document submitted
      tags: *a17
      security: *a18
  /kyc/documents/{id}:
    get:
      operationId: KycController_getDocument
      summary: Get KYC document by ID
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        "200":
          description: KYC document found
        "404":
          description: KYC document not found
      tags: *a17
      security: *a18
  /kyc/integrations/status:
    get:
      operationId: KycController_getIntegrationsStatus
      summary: Get KYC integration status (provider/storage)
      parameters: []
      responses:
        "200":
          description: KYC integration status
      tags: *a17
      security: *a18
  /kyc/submit:
    post:
      operationId: KycController_submit
      summary: Submit KYC information (legacy)
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitKycDto"
      responses:
        "200":
          description: KYC submitted
      tags: *a17
      security: *a18
  /kyc/documents/{id}/status:
    patch:
      operationId: KycController_updateDocumentStatus
      summary: Update KYC document status (admin only)
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateKycDocumentStatusDto"
      responses:
        "200":
          description: Document status updated
        "404":
          description: Document not found
      tags: *a17
      security: *a18
info:
  title: Digital Banking Core API
  description: API surface for the digital banking MVP core components.
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    RegisterDto:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    LoginDto:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    RefreshTokenDto:
      type: object
      properties:
        refreshToken:
          type: string
      required:
        - refreshToken
    OAuthTokenDto:
      type: object
      properties: {}
    CreateAccountDto:
      type: object
      properties:
        currency:
          type: string
          default: USD
          description: ISO currency code
        initialBalance:
          type: number
          default: 0
      required:
        - currency
    FundAccountDto:
      type: object
      properties: {}
...
    RegisterCardDto:
      type: object
      properties: {}
    TokenizeCardDto:
      type: object
      properties: {}
    FreezeDto:
      type: object
      properties: {}
    UpdateLimitsDto:
      type: object
      properties: {}
...
    SubmitKycDto:
      type: object
      properties: {}
    SubmitKycDocumentDto:
      type: object
      properties: {}
    UpdateKycDocumentStatusDto:
      type: object
      properties: {}
    UpdateKycStatusDto:
      type: object
      properties: {}